+++ lib/omniauth/amazon.rb
@@ -1,4 +1,0 @@
-# frozen_string_literal: true
-
-require 'omniauth/amazon/version'
-require 'omniauth/strategies/amazon'