+++ lib/omniauth/amazon_business.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+require 'omniauth/amazon_business/version'
+require 'omniauth/strategies/amazon_business'